requirements.txt
asgiref==3.5.2
Babel==2.11.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-adminlte-3==0.1.6
django-dotenv==1.4.2
django-mathfilters==1.0.0
django-pdfkit==0.3.1
django-phonenumber-field==7.0.1
idna==3.4
pdfkit==0.6.0
phonenumbers==8.13.3
psycopg2==2.9.4
py-moneyed==2.0
PyJWT==2.6.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.2
twilio==7.15.0
typing-extensions==4.4.0
urllib3==1.26.12
wkhtmltopdf==0.2
